extract parse_type from parser
